pyproject.toml
[project]
name = "eftx-antenna-pattern"
version = "0.1.0"
description = "EFTX antenna composition and ERP designer web app"
requires-python = ">=3.12"
dependencies = [
    "Flask>=3.0",
    "Flask-Login>=0.6.3",
    "Flask-WTF>=1.2.1",
    "Flask-Limiter>=3.5.1",
    "Flask-Mailman>=1.1.1",
    "Flask-SQLAlchemy>=3.1",
    "SQLAlchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "argon2-cffi>=23.1",
    "itsdangerous>=2.1",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "matplotlib>=3.9",
    "reportlab>=4.2",
    "pypdf>=5.0",
    "Pillow>=10.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5"
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
